Added CMake structure and install - builds, but has not been tested in other builds
